Add Join-able entries API to Storage
